refactor(build): use compile-time flake option
This makes the `nixos build` command more ergonomic to use if you
are a regular flake user. The compile-time build option -Dflake=[bool]
is true by default, so this is enabled for all users unless expliticly
opted out.

Legacy configuration users will need to build this package with
-Dflake=false in order to build their configurations. Legacy
configurations should not be prioritized equally to flakes,
as flakes are the new path forward, it seems.