feat(build): use FOD to build zig deps, clean up flake
I wanted to use a fixed-output derivation instead of using zon2nix
to do so, since this relied on a weird patch to get stuff working, and
zon2nix hasn't been touched in around 6 months despite having a bug
and an opened PR to fix it.

This also reworks some of the flake; instead of using gitignore.nix
and the Zig overlay, it uses the Zig package from nixpkgs and another
gitignore package from nixpkgs; the less deps, the better right?