chore(deps): remove zignix bindings usage, for now
It's just slowing development down for minimal gain. Plus, the value
of a static binary is not to be unstated.

Let's re-evaluate usage of this when we need it, now that we've used
it as a proof of concept. It's possible that we can just use the Nix
CLI as a means of getting all the information that we need for now;
yes, that does include options, which we can cache later.