fix: only preserve NIXOS_CLI_CONFIG in nix module
Oops. I didn't realize the implications of adding -E to indiscriminately
propagate the entire user environment. It seems to have propagated the
$USER var as well, so that's a no-go.

Just add it to the sudoers config in the module from now on, but we
must carefully take these things into consideration when implementing
things related to env vars from now on.