fix(option): disconnect table scrolling from starting row
This reworks some of the logic involved in rendering the option result
rows. Instead of reversing the sorting and rendering from top to bottom,
when the scrolling happens from bottom to top, this synchronizes the
two.

Also, this unpins the scrolling from the starting row. Now a user can
scroll without always moving the starting row.