feat: add option search UI (#35)
* feat(option): init option search UI skeleton

* feat(option): filter option list using query

* feat(option): use range highlighting to display results

* fix(option): disconnect table scrolling from starting row

This reworks some of the logic involved in rendering the option result
rows. Instead of reversing the sorting and rendering from top to bottom,
when the scrolling happens from bottom to top, this synchronizes the
two.

Also, this unpins the scrolling from the starting row. Now a user can
scroll without always moving the starting row.

* feat(option): allow configurable maximum distance score

* feat(option): wrap around when end indices are hit

* feat(option): show number of results in search box

* feat(option): show scrollable option preview

* docs: update roadmap

* feat(option): add loading message when building cache